Add long line support
We have really long lines in the config file sometimes because of
the ip whitelist. This just adds a patch to alter the defaults file
so the max line length is 10x it's default value. This might could
be done via an option to make insted, which would be cleaner.